update transition in subscribe ee (#149)
* Changed transitions based on spec.

* Fixed transition from Handshaking to HandshakeReconnecting
on HandshakeFailure in case Handshake has timetoken != 0

This situation happens when there is SubscriptionRestored on Unsubscribed state. In such case timetoken is associated with SubscriptionRestored event. Handshake call is executed with timetoken zero as a result new timetoken is returned. Next ReceiveMessages call is executed with timetoken provided in SubscriptionRestored event not returned by Handshake.